Only execute this if /run/ramalama exists

+++ container-images/scripts/build_llama_and_whisper.sh
@@ -255,7 +255,9 @@ clone_and_build_llama_cpp() {
 install_ramalama() {
   # link podman-remote to podman for use by RamaLama
   ln -sf /usr/bin/podman-remote /usr/bin/podman
-  python3 -m pip install /run/ramalama --prefix="$1"
+  if [ -e "/run/ramalama" ]; then
+    python3 -m pip install /run/ramalama --prefix="$1"
+  fi
 }
 
 main() {
